am62x: mcspi: fix MISRA.ETYPE.COMP.ASSIGN.2012
MISRA.ETYPE.COMP.ASSIGN.2012 - The value of a composite expression shall
not be assigned to an object with wider essential type.
Modify the type of the variable to fix the issue.

Fixes: SITSW-3820

+++ source/drivers/mcspi/v0/mcspi_v0.c
@@ -1466,19 +1466,19 @@ static void MCSPI_setChConfig(MCSPI_Config *config,
     {
         if(MCSPI_TR_MODE_TX_RX == chObj->chCfg.trMode)
         {
-            chObj->intrMask = ((CSL_MCSPI_IRQSTATUS_RX0_FULL_MASK      << (4U * chNum)) |
+            chObj->intrMask = (((uint32_t)CSL_MCSPI_IRQSTATUS_RX0_FULL_MASK << (4U * chNum)) |
                                (CSL_MCSPI_IRQSTATUS_TX0_EMPTY_MASK     << (4U * chNum)) |
                                (CSL_MCSPI_IRQSTATUS_TX0_UNDERFLOW_MASK << (4U * chNum)) |
                                (CSL_MCSPI_IRQSTATUS_RX0_OVERFLOW_MASK));
         }
         else if(MCSPI_TR_MODE_TX_ONLY == chObj->chCfg.trMode)
         {
-            chObj->intrMask = ((CSL_MCSPI_IRQSTATUS_TX0_EMPTY_MASK     << (4U * chNum)) |
+            chObj->intrMask = (((uint32_t)CSL_MCSPI_IRQSTATUS_TX0_EMPTY_MASK << (4U * chNum)) |
                                (CSL_MCSPI_IRQSTATUS_TX0_UNDERFLOW_MASK << (4U * chNum)));
         }
         else
         {
-            chObj->intrMask = ((CSL_MCSPI_IRQSTATUS_RX0_FULL_MASK     << (4U * chNum)) |
+            chObj->intrMask = (((uint32_t)CSL_MCSPI_IRQSTATUS_RX0_FULL_MASK << (4U * chNum)) |
                                (CSL_MCSPI_IRQSTATUS_RX0_OVERFLOW_MASK));
         }
     }
